+++ tests/SolSharp.Programs.Tests/TransactionTests.cs
@@ -22,6 +22,19 @@ public static class TransactionTests
 
     private static byte[] Fill(byte value) => [.. Enumerable.Repeat(value, PublicKey.Length)];
 
+    private sealed class TestSigner(PublicKey publicKey, byte[]? signature) : ISigner
+    {
+        public PublicKey PublicKey { get; } = publicKey;
+
+        public int CallCount { get; private set; }
+
+        byte[] ISigner.Sign(ReadOnlySpan<byte> message)
+        {
+            CallCount++;
+            return signature!;
+        }
+    }
+
     private static Transaction BuildTransfer(out Keypair payer)
     {
         payer = Keypair.FromSeed(Fill(1));
@@ -63,8 +76,9 @@ public static class TransactionTests
         {
             // Arrange
             var transaction = BuildTransfer(out var payer);
-            using var stranger = Keypair.Generate();
+            var stranger = Keypair.Generate();
             using (payer)
+            using (stranger)
             {
                 // Act
                 Action act = () => transaction.Sign(stranger);
@@ -72,6 +86,153 @@ public static class TransactionTests
                 // Assert
                 act.Should().Throw<ArgumentException>();
             }
+        }
+
+        [Test]
+        public void LaterNonRequiredSigner_IsRejectedBeforeAnySignerIsCalled()
+        {
+            // Arrange
+            var transaction = BuildTransfer(out var payer);
+            using (payer)
+            {
+                var firstSigner = new TestSigner(payer.PublicKey, new byte[Transaction.SignatureLength]);
+                var stranger = new TestSigner(new PublicKey(Fill(9)), new byte[Transaction.SignatureLength]);
+
+                // Act
+                Action act = () => transaction.Sign(firstSigner, stranger);
+
+                // Assert
+                act.Should().Throw<ArgumentException>();
+                firstSigner.CallCount.Should().Be(0);
+                stranger.CallCount.Should().Be(0);
+            }
+        }
+
+        [Test]
+        public void NullSignerElement_Throws()
+        {
+            // Arrange
+            var transaction = BuildTransfer(out var payer);
+            using (payer)
+            {
+                ISigner[] signers = [null!];
+
+                // Act
+                Action act = () => transaction.Sign(signers);
+
+                // Assert
+                act.Should().Throw<ArgumentNullException>().WithParameterName(nameof(signers));
+            }
+        }
+
+        [TestCase(63)]
+        [TestCase(65)]
+        public void InvalidSignatureLength_Throws(int length)
+        {
+            // Arrange
+            var transaction = BuildTransfer(out var payer);
+            using (payer)
+            {
+                var signer = new TestSigner(payer.PublicKey, new byte[length]);
+
+                // Act
+                Action act = () => transaction.Sign(signer);
+
+                // Assert
+                act.Should().Throw<ArgumentException>().WithMessage("*64-byte*");
+            }
+        }
+
+        [Test]
+        public void NullSignature_Throws()
+        {
+            // Arrange
+            var transaction = BuildTransfer(out var payer);
+            using (payer)
+            {
+                var signer = new TestSigner(payer.PublicKey, signature: null);
+
+                // Act
+                Action act = () => transaction.Sign(signer);
+
+                // Assert
+                act.Should().Throw<ArgumentException>().WithMessage("*64-byte*");
+            }
+        }
+
+        [Test]
+        public void LaterInvalidSignature_DoesNotCommitEarlierSignature()
+        {
+            // Arrange
+            var payer = new PublicKey(Fill(1));
+            var second = new PublicKey(Fill(2));
+            var instruction = new Instruction
+            {
+                ProgramId = new PublicKey(Fill(9)),
+                Accounts = [AccountMeta.ReadonlySigner(second)],
+                Data = [7]
+            };
+            var message = Message.Compile(payer, new PublicKey(Fill(8)).ToString(), [instruction]);
+            var transaction = Transaction.Create(message);
+            var firstSigner = new TestSigner(payer, [.. Enumerable.Repeat((byte)0xAB, Transaction.SignatureLength)]);
+            var invalidSecondSigner = new TestSigner(second, new byte[Transaction.SignatureLength - 1]);
+
+            // Act
+            Action act = () => transaction.Sign(firstSigner, invalidSecondSigner);
+
+            // Assert
+            act.Should().Throw<ArgumentException>();
+            transaction.Serialize().Skip(1).Take(2 * Transaction.SignatureLength).Should().OnlyContain(value => value == 0);
+        }
+
+        [Test]
+        public void MutatingReturnedSignature_DoesNotMutateTransaction()
+        {
+            // Arrange
+            var transaction = BuildTransfer(out var payer);
+            using (payer)
+            {
+                byte[] signature = [.. Enumerable.Repeat((byte)0xAB, Transaction.SignatureLength)];
+                var signer = new TestSigner(payer.PublicKey, signature);
+                transaction.Sign(signer);
+
+                // Act
+                signature[0] = 0;
+
+                // Assert
+                transaction.Serialize()[1].Should().Be(0xAB);
+            }
+        }
+
+        [Test]
+        public void SignerSlotsRemainBoundToCapturedMessageAfterAccountKeyMutation()
+        {
+            // Arrange
+            var payer = new PublicKey(Fill(1));
+            var second = new PublicKey(Fill(2));
+            var stranger = new PublicKey(Fill(3));
+            var instruction = new Instruction
+            {
+                ProgramId = new PublicKey(Fill(9)),
+                Accounts = [AccountMeta.ReadonlySigner(second)],
+                Data = [7]
+            };
+            var message = Message.Compile(payer, new PublicKey(Fill(8)).ToString(), [instruction]);
+            var transaction = Transaction.Create(message);
+            var signature = new byte[Transaction.SignatureLength];
+            var payerSigner = new TestSigner(payer, signature);
+            var secondSigner = new TestSigner(second, signature);
+            var strangerSigner = new TestSigner(stranger, signature);
+            transaction.Sign(payerSigner);
+
+            // Act
+            ((List<PublicKey>)message.AccountKeys)[1] = stranger;
+            Action strangerAct = () => transaction.Sign(strangerSigner);
+            Action originalSignerAct = () => transaction.Sign(secondSigner);
+
+            // Assert
+            strangerAct.Should().Throw<ArgumentException>();
+            originalSignerAct.Should().NotThrow();
         }
     }
 
@@ -91,6 +252,25 @@ public static class TransactionTests
                 // Assert
                 bytes[0].Should().Be(1); // ShortVec(1): one signature slot
                 bytes.Skip(1).Take(Transaction.SignatureLength).Should().OnlyContain(b => b == 0);
+            }
+        }
+
+        [Test]
+        public void SignedTransaction_IsStableAfterMessageMutation()
+        {
+            // Arrange
+            var transaction = BuildTransfer(out var payer);
+            using (payer)
+            {
+                transaction.Sign(payer);
+                var before = transaction.Serialize();
+
+                // Act
+                transaction.Message.Instructions[0].Data[0] ^= 0xFF;
+
+                // Assert
+                transaction.Serialize().Should().Equal(before);
+                transaction.GetSerializedLength().Should().Be(before.Length);
             }
         }
     }